requirements.txt
Flask==2.2.2
Flask-PyMongo==2.3.0
Werkzeug==2.3.1
